requirements.txt
PyQt5==5.15.10
Pillow==10.1.0
mss==9.0.1
opencv-python==4.9.0.80
numpy==1.26.4
